+++ Changelog.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file
 
 
 ## Release 1.1
+- FIX : Warning and create extrafields - *17/02/2026* - 1.1.2
 - FIX : Compat v23
   Changed Dolibarr compatibility range to 17 min - 23 max - *20/01/2026* - 1.1.1
 - FIX : Compat v20